requirements.txt
appdirs==1.4.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
cramjam==2.11.0
dotenv==0.9.9
et_xmlfile==2.0.0
fastparquet==2024.11.0
fsspec==2025.9.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
nfl_data_py==0.3.3
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==1.5.3
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
uritemplate==4.2.0
urllib3==2.5.0
